digital: Mostly whitespace: pull main work in correlate_and_sync_cc out of "if"
This is a mostly whitespace change, with one logic inversion and one use of
a continue statement, to move the main part of the work in
correlate_and_sync_cc_impl work() function out of an "if {}" clause and
move a trivial "else i++" clause.

This is in preparation for addtional changes.